fix(CoverageReport): update vacation_days field type and adjust related calculations
This commit changes the 'vacation_days' field in the CoverageReport model from float to integer, ensuring accurate data representation. The CoverageReportTable is updated to simplify the numeric formatting for vacation days. Additionally, the GetCoverageReportData SQL procedure is modified to cast the sum of vacation days as an unsigned integer, aligning with the new model definition and improving data integrity in coverage reports.

+++ app/Models/CoverageReport.php
@@ -47,7 +47,7 @@ class CoverageReport extends Model
         'actual_visits' => 'integer',
         'call_rate' => 'float',
         'total_visits' => 'integer',
-        'vacation_days' => 'float',
+        'vacation_days' => 'integer',
         'daily_report_no' => 'integer'
     ];
